Add Prometheus query as new monitor type (#14)
* Kinda working prometheus query

* Fix formatting: add newline for improved readability in runPromQuery function

* Enhance logging in monitor: use StatusString for better readability and add error handling for empty Prometheus query results

* Add Prometheus monitor support to documentation

* Remove obsolete PROM TEST monitor configuration

* Add header comments to Dockerfile.database for clarity

* Reduce default timeout to 5 seconds and streamline output handling in runPromQuery function

* Refactor runPromQuery: streamline context creation and add case for ValMatrix

* Update Prometheus monitor section in README: clarify description and enhance metric representation

* chore: update frontend dependencies and fix package.json overrides

- Bump versions of @azure/msal-browser, @azure/msal-react, FontAwesome, Bootstrap, React, and related packages in frontend/package.json.
- Update devDependencies including ESLint, TypeScript, and Vite to their latest versions.
- Add temporary override for MSAL compatibility with React 19.

chore: update Go module dependencies

- Upgrade Go version to 1.25.0 in go.mod.
- Update various dependencies including chi, prometheus/client_golang, and others to their latest versions.
- Adjust indirect dependencies accordingly.

docs: improve readme formatting

- Highlight default listening ports for API and frontend in readme.md.
- Fix formatting for source code links in readme.md.

chore: add initial package-lock.json

- Create package-lock.json for frontend dependencies.

* chore: update architecture diagram in draw.io format

* chore: remove package-lock.json file

+++ services/common/monitor/tcp.go
@@ -35,7 +35,7 @@ func (m *Monitor) runTCP() *result.Result {
 		return result.NewFailedResult(m.Name, m.Target, m.ID, err)
 	}
 
-	r.Value = int(time.Since(start).Milliseconds())
+	r.Value = float64(time.Since(start).Milliseconds())
 
 	outputs := map[string]any{
 		"respTime":  r.Value,
